feat: create moa - step page3 : layout (MOA-106)
