+++ src/app/[year]/(create)/create-moa/page.tsx
@@ -1,10 +1,10 @@
-import CreateMoaStep4 from "./step4"
+import CreateMoaStep3 from "./step3"
 
 export default function CreateMoaStep() {
   return (
     <>
       <h1>Create Moa - Step Page</h1>
-      <CreateMoaStep4></CreateMoaStep4>
+      <CreateMoaStep3></CreateMoaStep3>
     </>
   )
 }
